pyproject.toml
[project]
name = "lapua-kaupunki-rag"
version = "0.1.0"
description = "Lapuan kaupungin pöytäkirjoihin perustuva RAG-järjestelmä (Docling + BGE-M3 + Qdrant + agentit)."
requires-python = ">=3.10"
readme = "README.md"
license = { text = "MIT" }
authors = [
  { name = "Lapua RAG", email = "info@example.com" },
]

dependencies = [
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "docling>=2.0.0",
  "numpy>=1.26.0",
  "qdrant-client>=1.11.0",
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "httpx>=0.27.0",
  "networkx>=3.2.0",
  "python-dotenv>=1.0.0",
  "groq>=0.11.0",
]

[project.optional-dependencies]
embeddings = [
  "flagembedding>=1.2.10",  # BGE-M3
  "scipy>=1.11.0",
]

[tool.setuptools]
package-dir = {"" = "packages"}

[tool.setuptools.packages.find]
where = ["packages"]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true

[tool.ruff]
line-length = 100

[tool.black]
line-length = 100
target-version = ["py310"]


